CP-54351 move devscan from /opt
